requirements.txt
pyyaml

#To install: pip install -r requirements.txt
